+++ nimare/dataset.py
@@ -12,8 +12,8 @@ import pandas as pd
 import nibabel as nib
 
 from .base import NiMAREBase
-from .utils import (tal2mni, mni2tal, mm2vox, get_template, listify,
-                    try_prepend, find_stem, get_masker)
+from .transforms import tal2mni, mni2tal, mm2vox
+from .utils import get_template, listify, try_prepend, find_stem, get_masker
 
 LGR = logging.getLogger(__name__)
