+++ nimare/meta/cbma/kernel.py
@@ -10,7 +10,7 @@ import nibabel as nib
 
 from nilearn.image import resample_to_img, math_img
 from .utils import compute_ma, get_ale_kernel, peaks2maps
-from ...utils import vox2mm
+from ...transforms import vox2mm
 
 from ...base import Transformer
