+++ nimare/meta/cbma/mkda.py
@@ -10,10 +10,9 @@ from tqdm.auto import tqdm
 from scipy import ndimage, special
 from statsmodels.sandbox.stats.multicomp import multipletests
 
-from .kernel import MKDAKernel, KDAKernel
+from .kernel import MKDAKernel, KDAKernel, KernelTransformer
 from ...results import MetaResult
 from ...base import CBMAEstimator
-from .kernel import KernelTransformer
 from ...stats import null_to_p, one_way, two_way
 from ...transforms import p_to_z
 from ...due import due
